Move hunk allocation functions to their own file
Also rename Com_InitHunkMemory to Hunk_Init
